+++ plone/app/contenttypes/tests/test_collection.py
@@ -6,6 +6,7 @@ import unittest2 as unittest
 
 from zope.component import createObject
 from zope.component import queryUtility
+from zope.interface import alsoProvides
 
 from transaction import commit
 
@@ -23,6 +24,9 @@ from plone.app.testing import TEST_USER_ID, TEST_USER_NAME, \
 from plone.app.contenttypes.interfaces import ICollection
 from plone.app.contenttypes.behaviors.collection import ICollection as \
     ICollection_behavior
+
+from plone.app.layout.navigation.interfaces import INavigationRoot
+
 
 query = [{
     'i': 'Title',
@@ -336,6 +340,47 @@ class PloneAppCollectionViewsIntegrationTest(unittest.TestCase):
         wrapped.query = query
         imagecount = wrapped.getFoldersAndImages()['total_number_of_images']
         self.assertEqual(imagecount, 2)
+
+    def test_respect_navigation_root(self):
+        portal = self.layer['portal']
+        login(portal, 'admin')
+
+        # Create two subsites i.e create two folders and mark them with
+        # INavigationRoot
+        for i in xrange(1, 3):
+            folder_id = 'folder{}'.format(i)
+            portal.invokeFactory('Folder',
+                                 folder_id,
+                                 title='Folder{}'.format(i))
+            folder = portal[folder_id]
+            alsoProvides(folder, INavigationRoot)
+        folders = (portal['folder1'], portal['folder2'])
+
+        # Add a content item to each folder
+        for f in folders:
+            f_id = f.getId()
+            f.invokeFactory('Document',
+                            'item_in_{}'.format(f_id),
+                            title='Item In {}'.format(f_id))
+
+        # Add a collection to folder1
+        folder1 = folders[0]
+        folder1.invokeFactory('Collection',
+                              'collection1',
+                              title='Collection 1')
+        collection1 = folder1['collection1']
+        wrapped = ICollection_behavior(collection1)
+        wrapped.query = [{
+            'i': 'portal_type',
+            'o': 'plone.app.querystring.operation.string.is',
+            'v': 'Document',
+        }]
+
+        # Check if only the item inside folder1 is returned, since it's a
+        # navigation root.
+        items = wrapped.results(batch=False)
+        ids = [i.getId() for i in items]
+        self.assertListEqual(ids, ['item_in_folder1'])
 
 
 class PloneAppCollectionEditViewsIntegrationTest(unittest.TestCase):